Update Mono-sqlite to track the RemoveGlobalSystems branch

+++ EDDiscovery/DB/InMemory/VisitedSystemsClass.cs
@@ -9,11 +9,11 @@ namespace EDDiscovery2.DB.InMemory
     {
         // For when you need a minimal version and don't want to mess up the database. 
         // Useful for creation of test doubles
-        public int id { get; set; }
+        public long id { get; set; }
         public string Name { get; set; }
         public DateTime Time { get; set; }
         public int Commander { get; set; }
-        public int Source { get; set; }
+        public long Source { get; set; }
         public string Unit { get; set; }
         public bool EDSM_sync { get; set; }
         public int MapColour { get; set; }
